Complete redo as a pretty application/crate
Prior to this we had some minimal code for enumerating tags/releases
with absolutely no introspection capabilites. Before that rot set in
we've overhauled now to use tracing and color-eyre for full insight
as well as having a main entry point for "playground" experiments.

The GitHub plugin now merges the releases/tags data into a single unified
representation. The plan of course is to cache much of this information
and abstract across various backend impls, allowing a wide variety of
use cases:

 - Upstream release checking
 - Generation of git commits with release notes
 - Discovery of signatures for tarball URIs (gpg verify at build)
 - and more. :)

Signed-off-by: Ikey Doherty <ikey@aerynos.com>